Make login page responsive (#229)
Closes #94 
---
**Updates after submitting PR #229** | 5/9/24
> Please look at adding a simple test to ensure the page has the
elements we're looking for. Then, use storybook visual snapshots to
ensure mobile and desktop views are correct.
- [x] Add unit tests with RTL
- [x] Add E2E test with Playwright
- [ ] ~~Add Storybook visual testing~~ - Currently blocked by needing to
connect Chromatic to GitHub and granting access to the project. Should
we all create an account and request access or do we use a team-wide
account?
---
## Demo Video 👇🏻

https://github.com/LetsGetTechnical/gridiron-survivor/assets/40150036/205c3942-bbba-4df5-bb6e-492e2dca5602
## Changes
- Changed `img` rule in `globals.css` to `.weeklyPickImage` so it does
not affect all images on the app
- Fixed some styles on the root `layout.tsx` and the login `layout.tsx`
- Our theming was never truly implemented and, as such, more changes are
needed. For now, I made it so the layout `body` respects the theming
currently in place and applies the class of `dark` when dark mode is on.
This, along with using our CSS variables, will help with the light and
dark mode theming.
  - More work needs to be done on this.
- Added `layout.tsx` inside login directory
- Reworked login `page.tsx` from the ground up to make it fully
responsive.
  - Reworked from the ground up to go from mobile --> desktop
- Added tailwind classes for width and height in the `Logo` component to
avoid using tailwind arbitrary values
- Added `weeklyPickImage` class to the `weeklyPickImage` component image
for more specific styling
- Add RTL and Jest testing for login page components with Shashi's help
- Add Playwright E2E for UX for login page