refine theme and colors (#505)
Closes #372 
Closes #506 

Development of the front-end started with both a light and dark theme,
as such, appropriate tailwind classes were added to allow for dynamic
switching. However, as time has went on it seems the light theme has
been abandoned. I feel we should decide what we are targeting, rework
the existing code accordingly, and ensure all team members are on the
same page.

- [X] Remove all standalone theme options
- [X] Ensure all styling is done using CSS Variables unless it is an
appropriate one-off use case
- [X] Ensure Dark theme CSS variables are aligned with the Figma design

CSS Variables should be used wherever possible and are preferred.

# Video Demo
**In the demo, you can see the white border of my browser indicating it
is in light mode. In dark mode, it has a dark indigo border as seen
below. the video**


https://github.com/user-attachments/assets/1e90849e-99fa-4153-9e33-044ed68b0c6e

![Arc Space
1-000261](https://github.com/user-attachments/assets/e164fbb5-6829-4cf8-9cf5-14cc1644d755)