Alex/add redirect for team selections (#519)
Closes #515

Redirect user to the league home after they make a successful pick.
Allowing users to make selections to their Entries faster.

- [x] Added router.push to the handleWeeklyPickChange function to
redirect the user back to their entry page.
- [x] For testing I had to grab the 'team-radio' test id and had to use
"findAll" because I had to await it to simulate the page loading and
there are multiple team radio's.
- [x] Then in that same test, I had to put selectedTeams[0] to simulate
me clicking the first team option. The [0] is the index.

https://github.com/user-attachments/assets/3ef29b76-1b93-4e0d-8cef-f010964ba4de