Danielle/homepage-flickers-as-it-checks-authentication (#654)
Closes #635 

Issue:
If user is logged in and goes to gridironsurvivor.com or
gridironsurvivor.com/login, the login page displays briefly before the
user is redirected to the leagues landing page.

Desired behavior:
A user who is already logged in should see the global spinner (pulsing
logo) while authentication takes place, then is taken straight to
leagues landing page. Login page should not display at all.
For a user who is not logged in, the global spinner should display while
authentication is checked, then see the login page.

Solution:

- updated the `isSignedIn` variable to have 3 states instead of the
previous 2 - null, true, false. Null will satisfy the condition where
authentication is being checked. If the user is logged in, state goes to
true. If not logged in, state goes to false.
- wrapped the page in the `GlobalSpinner` and added conditional
rendering so that, if the user is logged in, they only see the spinner
followed by the page redirect
- updated the tests to reflect the spinner displaying while
authentication is checked, then either redirecting the logged in user to
the leagues landing page or loading the login page if the user is not
logged in