Alex/implement admin league operation (#489)
Closes #480 

Shashi requested that I only do these things:

- [x] Render the table
- [x] Get the specific data for this table and show it

Let me break down what I did and why.

1. getAllLeagueEntries(); apiFunction:
- in order to get the numbered data for the entries, I had to create an
apiFunction to get this and return the length of entries.
- I set the Query.limit to 5000 because appwrite only returns 25 by
default (there was more than 25 to bring back).
- the reason it is filtering is because I can't actually query 'league'
in the database because it is considered a relationship. So, I had to
filter and grab entry.league(the relationship) that is matched with the
leagueId that I pass in.

2. For the page.tsx the `fetchData` is what is needed to grab certain
data to pass into the table to be rendered.
- the first function gets the number of total entries and entries that
aren't eliminated
- the second function gets info from all leagues the user is a part of
- the `combinedData` is there because I was getting data from 2 separate
collections in the database. Then, I HAD to combine them into 1 to be
able to pass into the table's data={}.

3. The IEntryWithLeague props extends ILeague because it uses almost all
props in ILeague
- then I created 2 new props to let it know that I will be receiving
them also into the table
- then I am able to define the table column header as `IEntryWithLeague`
because it uses all those props

4. The cell: {row} grabs the props by their names, so it knows what data
it is grabbing to display its value.
- because of fetchData, the table columns works closely with the data
props and knows what to display

5. the accessorKey is the actual identity of the prop it is grabbing

6. Some will have `as number` at the end, to handle type errors and let
the table know that I am specifically rendering a number value.

7. The reason for `sortingFn` for some columns and not all is because
originally survivors and participants were a string[] and I grabbed the
.length of them. So to accurately sort i had to add the sorting code in.



https://github.com/user-attachments/assets/b066b505-0c08-43ad-8631-2fb973d1b52f

---------

Co-authored-by: Chris Nowicki <102450568+chris-nowicki@users.noreply.github.com>